pyproject.toml
[project]
name = "o-positive"
version = "0.1.0"
description = "Workout Schedule Generator API with LLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.123.5",
    "langchain>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.32.1",
]
